Datastore, more hooks, and better typing (#2)
* wip: datastore and hooks
* wip: construct the ColumnDefinition appropriately0
* wip: support virtual attributes
* chore: standard

Datastore and hooks

Add support for datastore as simple KV store to avoid any overriding
initialization logic in the importer. Support this functionaity with a
before_import_blocks hooks that allows for arbitrary code to be run
prior to persistence, which is particularly useful for preloads
Virtual attributes

Adds support for virtual columns using the virtual: true. These are
columns that should be read in, but shouldn't actually be set are the
model. This allows them to easily be referenced in the import, without
actually attempting to set attributes on the model that don't exist.
Note that this is distinct from a Rails virtual attribute in
ActiveModel, and its name is thus subject to change
Typing

Improve typing by creating ConfigInterface and using that to add full
support for typing to the application with coverage for Dsl as well. A
few methods are still left without clear typing / documentation, as the
signature continues to be improved for those

Appropriate specs are added to ensure this new functionality works as
expected